cdk-lambda/requirements.txt
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_s3
aws-cdk.aws_dynamodb
aws-cdk.aws_events_targets
aws-cdk.aws_s3_notifications